Adding Base Service & Item's complete implementation

+++ Backend Services/src/main/java/com/restaurant/backend/payloads/ItemDTO.java
@@ -19,5 +19,5 @@ public class ItemDTO extends BaseDTO{
     private double price;
 
     // takes only item category id;
-    private Integer itemCategory;
+    private long itemCategory;
 }
